# vectorized loads/stores