# Explanation

If we can tell the CUDA compiler that our memory is aligned, it can emit vectorized load instructions to load four floating point values at once. This code is a basic implementation of this idea; it creates a new codepath when alignment by four `weight_t`s is available. This is basically just a proof of concept, and should be cleaned up a little more before merging if you are interested.

# Speedup

This looks to give about a 7% speedup in my crude benchmark on an A100. To benchmark, I ran a forward scan on a `[8, 1536, 65536]` tensor. I got the following:

```
Old code:
12.4 ms ± 695 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)

New code:
11.4 ms ± 742 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

Code for benchmark was just this, in a notebook:

```python
from accelerated_scan.ref import scan as scan_ref
from accelerated_scan.warp import scan_forward as scan_warp
import torch

batch_size, dim, seqlen = 8, 1536, 65536
gates = 0.999 + 0.001 * torch.rand(batch_size, dim, seqlen, device="cuda")
tokens = torch.rand(batch_size, dim, seqlen, device="cuda")

def warp_fn():
    scan_warp(gates, tokens).sum().item()

%timeit warp_fn()
```

# Testing

I tested with this code in a notebook:

```python
from accelerated_scan.ref import scan as scan_ref
from accelerated_scan.warp import scan_forward as scan_warp
import torch

batch_size, dim, seqlen = 3, 1536, 4096
gates = 0.999 + 0.001 * torch.rand(batch_size, dim, seqlen, device="cuda")
tokens = torch.rand(batch_size, dim, seqlen, device="cuda")

for reverse in [False, True]:
    out_ref = scan_ref(gates, tokens, reverse=reverse)
    out_warp = scan_warp(gates, tokens, reverse=reverse)
    print((out_ref - out_warp).abs().max().item())
```

I did not yet modify the unit tests to check that the misaligned code path has no regressions, since the current tests will likely not trigger it. We should modify the tests to also explicitly _misalign_ the inputs before applying the scan.