@proger 2024-03-13
Thanks @unixpickle, I love it! Is there any reason not to make this always a default? Going to test on some 3090s and 4090s in a bit.