@unixpickle 2024-03-13
@proger I think it could definitely be the default. Occasionally users may want to operate on unaligned tensors, though, and I figured we might not want to get rid of that option.

One idea I had to make this implementation simpler and more general: replace `AlignedQuad` with something more like `AlignedTuple<N, weight_t>` and `UnalignedTuple<N, weight_t>` where N is just `kNStepsPerThread`. This way, we could always use the same implementation for all tuple sizes (not just 4, but also 2 and 1), and the caller could toggle between `AlignedTuple` and `UnalignedTuple` depending on the alignment of the data. If I get time today, I can draft up this implementation and either push it here or make a new PR.

Very curious what your benchmarking and testing yields! I hope this speedup ends up being worth it :)