@unixpickle 2024-03-13
Just pushed an update to implement the idea I described above. Now there's less code overall, since aligned/unaligned code paths are shared. Somehow it seems like the unaligned code path might also be a bit faster than the old code (not sure why).