@proger 2024-03-14
This looks awesome! Waiting for some 4090 time to test