@proger 2024-03-15
Running `tests/bench.py`, available from `git checkout main -- tests/bench.py`:

4090 before, CUDA 12.2:

```
accelerated_scan: forward speed of (8,1536,seqlen), inference mode:
   SEQUENCE_LENGTH       warp
0            128.0   0.028879
1            256.0   0.052529
2            512.0   0.096415
3           1024.0   0.178424
4           2048.0   0.340968
5           4096.0   0.669595
6           8192.0   1.318381
7          16384.0   2.611871
8          32768.0   5.199120
9          65536.0  10.377000

accelerated_scan: reverse speed of (8,1536,seqlen), inference mode:
   SEQUENCE_LENGTH       warp
0            128.0   0.030199
1            256.0   0.052682
2            512.0   0.095039
3           1024.0   0.177528
4           2048.0   0.342718
5           4096.0   0.670403
6           8192.0   1.318333
7          16384.0   2.611830
8          32768.0   5.198706
9          65536.0  10.374404
```

After:

```
accelerated_scan: forward speed of (8,1536,seqlen), inference mode:
   SEQUENCE_LENGTH       warp
0            128.0   0.028585
1            256.0   0.052536
2            512.0   0.096318
3           1024.0   0.177983
4           2048.0   0.339779
5           4096.0   0.662888
6           8192.0   1.319535
7          16384.0   2.618453
8          32768.0   5.226898
9          65536.0  10.446727

accelerated_scan: reverse speed of (8,1536,seqlen), inference mode:
   SEQUENCE_LENGTH       warp
0            128.0   0.030207
1            256.0   0.052668
2            512.0   0.094981
3           1024.0   0.176923
4           2048.0   0.342622
5           4096.0   0.662547
6           8192.0   1.319268
7          16384.0   2.619458
8          32768.0   5.223196
9          65536.0  10.427466
```

3090, CUDA 12.1, before. Reverse is noticeably slower than forward.

```
accelerated_scan: forward speed of (8,1536,seqlen), inference mode:
   SEQUENCE_LENGTH       warp
0            128.0   0.030507
1            256.0   0.053015
2            512.0   0.106897
3           1024.0   0.190742
4           2048.0   0.395175
5           4096.0   0.828499
6           8192.0   1.752365
7          16384.0   3.568564
8          32768.0   7.222955
9          65536.0  14.608726

accelerated_scan: reverse speed of (8,1536,seqlen), inference mode:
   SEQUENCE_LENGTH       warp
0            128.0   0.028869
1            256.0   0.052325
2            512.0   0.104420
3           1024.0   0.192586
4           2048.0   0.398766
5           4096.0   0.901066
6           8192.0   1.837662
7          16384.0   3.721098
8          32768.0   7.512320
9          65536.0  15.305217
```

After: 

```
accelerated_scan: forward speed of (8,1536,seqlen), inference mode:
   SEQUENCE_LENGTH       warp
0            128.0   0.030541
1            256.0   0.053536
2            512.0   0.106409
3           1024.0   0.190483
4           2048.0   0.386041
5           4096.0   0.726554
6           8192.0   1.471165
7          16384.0   2.998701
8          32768.0   6.066624
9          65536.0  12.241628

accelerated_scan: reverse speed of (8,1536,seqlen), inference mode:
   SEQUENCE_LENGTH       warp
0            128.0   0.029124
1            256.0   0.053477
2            512.0   0.104680
3           1024.0   0.192415
4           2048.0   0.399936
5           4096.0   0.746641
6           8192.0   1.538295
7          16384.0   3.139386
8          32768.0   6.246253
9          65536.0  12.700525
```


That's quite a boost for my Ampere/3090 setup! No noticeable effect on Ada/4090 though.